Python

+++ Actividades condicionales/Ejercicio 21 UD1.py
@@ -1,9 +1,9 @@
 # 21. Programa que calcula una ecuación de segundo grado. Controla que el valor de la raíz cuadrada no de un valor negativo
 import math
 
-valor_ax2 = int(input("Introduce el valor ax elevado ^ 2"))
-valor_bx = int(input("Introduce el valor bx"))
-valor_c = int(input("Introduce c"))
+valor_ax2 = float(input("Introduce el valor ax elevado ^ 2"))
+valor_bx = float(input("Introduce el valor bx"))
+valor_c = float(input("Introduce c"))
 
 discriminante = valor_bx ** 2 - (4 * valor_ax2 * valor_c)
